Fix dev mode

+++ packages/core/strapi/lib/commands/builders/typescript.js
@@ -5,10 +5,15 @@ const fs = require('fs-extra');
 const tsUtils = require('@strapi/typescript-utils');
 
 const cleanupDistDirectory = async distDir => {
+  if (!await fs.pathExists(distDir)){
+    return 
+  }
+
   const dirContent = await fs.readdir(distDir);
   const validFilenames = dirContent
     // Ignore the admin build folder
     .filter(filename => filename !== 'build');
+
 
   for (const filename of validFilenames) {
     await fs.remove(path.resolve(distDir, filename));
@@ -21,6 +26,8 @@ module.exports = async ({ srcDir, distDir, watch = false }) => {
   if (!isTSProject) {
     throw new Error(`tsconfig file not found in ${srcDir}`);
   }
+
+  
 
   await cleanupDistDirectory(distDir);
